refactor(#9): Change filename and function names
Changed index.ts to firebase.ts to better describe the API file.
Also changed fetching and adding to use await instead of .then()
